feat: (피드백 반영) 문자를 입력받아서 숫자로 컨버팅하는 로직 controller 로 이동
관련 링크
[https://github.com/next-step/java-lotto-pro/pull/802/files#r1008881038](https://github.com/next-step/java-lotto-pro/pull/802/files#r1008881038)
